Move UserSettingsScreen into self-contained directory
