+++ app/(user)/index.tsx
@@ -1,4 +1,4 @@
-import UserSettingsScreen from '@/src/components/screens/UserSettingsScreen';
+import UserSettingsScreen from '@/src/components/screens/user-settings/UserSettingsScreen';
 
 export default function Index() {
   return <UserSettingsScreen />;
